scripts/python-seed/requirements.txt
cassandra-driver==3.29.0
Faker==24.0.0